@@ -4,7 +4,7 @@ namespace FLEA\Acl\Table;
 
 
 /**
- * 定义 \FLEA\Acl_Table_UserGroupsHasRoles 类
+ * 定义 \FLEA\Acl\Table\UserGroupsHasRoles 类
  *
  * @author toohamster
  * @package Core
@@ -13,7 +13,7 @@ namespace FLEA\Acl\Table;
 
 
 /**
- * \FLEA\Acl_Table_UserGroupsHasRoles 用于关联用户组和角色
+ * \FLEA\Acl\Table\UserGroupsHasRoles 用于关联用户组和角色
  *
  * @package Core
  * @author toohamster
@@ -33,6 +33,6 @@ class UserGroupsHasRoles extends \FLEA\Db\TableDataGateway
      *
      * @var string
      */
-    public $tableName = 'user_groups_has_roles';
+    public string $tableName = 'user_groups_has_roles';
 
 }
