@@ -2,19 +2,19 @@
 
 require('FLEA.php');
 
-$dbDSN = array(
+$dbDSN = [
     'driver' => 'mysql',
     'host' => 'localhost',
     'login' => 'root',
     'password' => '',
     'database' => 'test'
-);
+];
 
 FLEA::setAppInf('dbDSN', $dbDSN);
 FLEA::setAppInf('internalCacheDir', 'D:/temp');
 
 $acl = FLEA::getSingleton('FLEA_Acl_Manager');
-/* @var $acl FLEA_Acl_Manager */
+/* @var $acl \FLEA\Acl\Manager */
 
-$user = $acl->getUserWithPermissions(array('username' => 'liaoyulei'));
+$user = $acl->getUserWithPermissions(['username' => 'liaoyulei']);
 dump($user);
