@@ -40,19 +40,17 @@ class File
      * @param array $struct
      * @param string $name
      * @param int $ix
-     *
-     * @return \FLEA\Helper\FileUploader_File
      */
     public function __construct($struct, $name, $ix = false)
     {
         if ($ix !== false) {
-            $s = array(
+            $s = [
                 'name' => $struct['name'][$ix],
                 'type' => $struct['type'][$ix],
                 'tmp_name' => $struct['tmp_name'][$ix],
                 'error' => $struct['error'][$ix],
                 'size' => $struct['size'][$ix],
-            );
+            ];
             $this->_file = $s;
         } else {
             $this->_file = $struct;
@@ -211,7 +209,7 @@ class File
             } elseif (strpos($allowExts, '|')) {
                 $exts = explode('|', $allowExts);
             } else {
-                $exts = array($allowExts);
+                $exts = [$allowExts];
             }
 
             $filename = $this->getFilename();
