@@ -38,8 +38,6 @@ class Image
      * 静态方法创建一个 \FLEA\Helper\Image 类的实例。
      *
      * @param resource $handle
-     *
-     * @return \FLEA\Helper\Image
      */
     public function __construct($handle)
     {
@@ -67,12 +65,12 @@ class Image
             $fileext = pathinfo($filename, PATHINFO_EXTENSION);
         }
         $fileext = strtolower($fileext);
-        $ext2functions = array(
+        $ext2functions = [
             'jpg' => 'imagecreatefromjpeg',
             'jpeg' => 'imagecreatefromjpeg',
             'png' => 'imagecreatefrompng',
             'gif' => 'imagecreatefromgif',
-        );
+        ];
         if (!isset($ext2functions[$fileext])) {
             throw new \FLEA\Exception_NotImplemented('imagecreatefrom' . $fileext);
         }
@@ -167,7 +165,7 @@ class Image
             $oy = ($height - $sy) / 2;
         }
 
-        list($r, $g, $b) = $this->extractColor($bgcolor, '0xffffff');
+        [$r, $g, $b] = $this->extractColor($bgcolor, '0xffffff');
         $bgcolor = imagecolorallocate($dest, $r, $g, $b);
         imagefilledrectangle($dest, 0, 0, $width, $height, $bgcolor);
         imagecolordeallocate($dest, $bgcolor);
@@ -195,14 +193,14 @@ class Image
 
         if (!is_array($nocut)) {
             if ($nocut) {
-                $nocut = array('enabled' => true, 'pos' => 'center', 'bgcolor' => '0xffffff');
+                $nocut = ['enabled' => true, 'pos' => 'center', 'bgcolor' => '0xffffff'];
             } else {
-                $nocut = array('enabled' => false);
+                $nocut = ['enabled' => false];
             }
         } else {
-            $nocut['enabled'] = isset($nocut['enabled']) ? $nocut['enabled']: true;
-            $nocut['pos'] = isset($nocut['pos']) ? $nocut['pos']: 'center';
-            $nocut['bgcolor'] = isset($nocut['bgcolor']) ? $nocut['bgcolor']: '0xffffff';
+            $nocut['enabled'] = $nocut['enabled'] ?? true;
+            $nocut['pos'] = $nocut['pos'] ?? 'center';
+            $nocut['bgcolor'] = $nocut['bgcolor'] ?? '0xffffff';
         }
 
         if ($nocut['enabled']) {
@@ -251,12 +249,12 @@ class Image
                 $oy = ($height - $sy * $ratio) / 2;
             }
 
-            list($r, $g, $b) = $this->extractColor($nocut['bgcolor'], '0xffffff');
+            [$r, $g, $b] = $this->extractColor($nocut['bgcolor'], '0xffffff');
             $bgcolor = imagecolorallocate($dest, $r, $g, $b);
             imagefilledrectangle($dest, 0, 0, $width, $height, $bgcolor);
             imagecolordeallocate($dest, $bgcolor);
 
-            $args = array($dest, $this->_handle, $ox, $oy, 0, 0, $dx, $dy, $sx, $sy);
+            $args = [$dest, $this->_handle, $ox, $oy, 0, 0, $dx, $dy, $sx, $sy];
         } else {
             // 允许图像溢出
             if ($sy * $ratio < $height) {
@@ -269,7 +267,7 @@ class Image
                 $sy = $height * $ratio;
             }
 
-            $args = array($dest, $this->_handle, 0, 0, 0, 0, $width, $height, $sx, $sy);
+            $args = [$dest, $this->_handle, 0, 0, 0, 0, $width, $height, $sx, $sy];
         }
 
         if ($highQuality) {
@@ -340,6 +338,6 @@ class Image
             $hex = $default;
         }
         $dec = hexdec($hex);
-        return array(($dec >> 16) & 0xff, ($dec >> 8) & 0xff, $dec & 0xff);
+        return [($dec >> 16) & 0xff, ($dec >> 8) & 0xff, $dec & 0xff];
     }
 }
