@@ -35,7 +35,7 @@ class RolesManager extends \FLEA\Db\TableDataGateway
      *
      * @var string
      */
-    public $tableName = 'roles';
+    public string $tableName = 'roles';
 
     /**
      * 角色名字段
@@ -48,8 +48,6 @@ class RolesManager extends \FLEA\Db\TableDataGateway
      * 构造函数
      *
      * @param array $params
-     *
-     * @return \FLEA\Rbac_RolesManager
      */
     public function __construct(?array $params = null)
     {
