@@ -9,7 +9,7 @@
  * @version $Id: ErrorMessage.php 972 2007-10-09 20:56:54Z qeeyuan $
  */
 
-return array(
+return [
     // FLEA_Exception_ExpectedFile
     0x0102001 => '需要的文件 "%s" 没有找到.',
     // FLEA_Exception_ExpectedClass
@@ -78,4 +78,4 @@ return array(
     // FLEA_Db_Exception_MissingLinkOption
     0x0202002 => '创建数据表关联时缺少必须的 "%s" 选项.',
 
-);
+];
